requirements.txt
openai
PyMuPDF
python-dotenv
chainlit
transformers
sentence-transformers
langchain
langchain-openai
langchain-community
langchain-huggingface
faiss-cpu
